maude.py: Extract maude related functions into own library, check maude version in test.py
This means that we only check the maude version once per test run, rather than each time we generate proofs.
